stories_microservices_post_service/requirements.txt
alembic==1.5.8
attrs==20.3.0
click==7.1.2
flasgger==0.9.5
Flask==1.1.2
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.2
mistune==0.8.4
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.4.7
text-unidecode==1.3
toml==0.10.2
Werkzeug==1.0.1
